Consolidate OKX client module
