+++ quant_intraday/exchange/okx_client.py
@@ -1,67 +1,76 @@
-import os, json, time, base64, hashlib, hmac, httpx
+from datetime import datetime, timezone
+import os, time, hmac, hashlib, base64, httpx, json
 
-OKX_REST = "https://www.okx.com"
+def _iso_from_ms(ms: int) -> str:
+    return datetime.fromtimestamp(ms/1000, tz=timezone.utc).isoformat(timespec="milliseconds").replace("+00:00","Z")
 
-def okx_sign(ts: str, method: str, path: str, body: str, secret: str) -> str:
-    msg = f"{ts}{method}{path}{body}".encode()
-    mac = hmac.new(secret.encode(), msg, hashlib.sha256).digest()
-    return base64.b64encode(mac).decode()
+def okx_sign(secret: str, prehash: str) -> str:
+    return base64.b64encode(hmac.new(secret.encode(), prehash.encode(), hashlib.sha256).digest()).decode()
 
 class OKXClient:
-    def __init__(self, key, secret, passphrase, account="trade", timeout=10, simulated: bool | None = None):
-        self.key, self.secret, self.passphrase = key, secret, passphrase
-        self.account = account
-        self.simulated = simulated if simulated is not None else (os.getenv("OKX_SIMULATED","0") == "1")
-        self.rest = httpx.Client(base_url=OKX_REST, timeout=timeout)
+    def __init__(self, key, secret, passphrase, account="trade", base_url="https://www.okx.com"):
+        self.key, self.secret, self.passphrase, self.account = key, secret, passphrase, account
+        self.base_url = base_url
+        self.rest = httpx.Client(base_url=base_url, timeout=10.0)
+        self._drift_ms = 0
+        self._drift_at = 0
 
-    def _headers(self, method, path, body):
-        ts = str(int(time.time()))
-        sign = okx_sign(ts, method, path, body, self.secret)
-        h = {
+    def server_time_ms(self) -> int:
+        r = self.rest.get("/api/v5/public/time", timeout=5.0)
+        r.raise_for_status()
+        return int(r.json()["data"][0]["ts"])
+
+    def _iso_ts(self) -> str:
+        try:
+            return _iso_from_ms(self.server_time_ms())
+        except Exception:
+            now = int(time.time()*1000) + getattr(self, "_drift_ms", 0)
+            return _iso_from_ms(now)
+
+    def _headers(self, method: str, path: str, body: str):
+        ts = self._iso_ts()
+        prehash = f"{ts}{method.upper()}{path}{body or ''}"
+        sign = okx_sign(self.secret, prehash)
+        hdr = {
+            "Content-Type": "application/json",
             "OK-ACCESS-KEY": self.key,
             "OK-ACCESS-SIGN": sign,
             "OK-ACCESS-TIMESTAMP": ts,
             "OK-ACCESS-PASSPHRASE": self.passphrase,
-            "Content-Type": "application/json"
         }
-        # prefer official simulated header; fallback env switch OKX_ACCOUNT=paper is still read by some legacy stacks
-        if self.simulated or os.getenv("OKX_ACCOUNT","").lower() in ("paper","demo","sim"):
-            h["x-simulated-trading"] = "1"
-        return h
-
-    def get(self, path, params=None):
-        r = self.rest.get(path, headers=self._headers("GET", path, ""), params=params or {})
-        r.raise_for_status(); return r.json()
-
-    def post(self, path, payload: dict):
-        body = json.dumps(payload)
-        r = self.rest.post(path, headers=self._headers("POST", path, body), content=body)
-        r.raise_for_status(); return r.json()
-
-    # ---- Convenience ----
-    def get_balance(self, ccy="USDT"):
-        j = self.get("/api/v5/account/balance")
-        data = j.get("data",[{}])[0]
-        for d in data.get("details", []):
-            if d.get("ccy") == ccy: return float(d.get("eq", 0))
-        return float(data.get("totalEq", 0))
+        if os.getenv("OKX_SIMULATED","0") in ("1","true","True"):
+            hdr["x-simulated-trading"] = "1"
+        if os.getenv("QI_DEBUG_HTTP","0") not in ("0","false","False",""):
+            safe = {k: ("***" if k.startswith("OK-ACCESS-") else v) for k,v in hdr.items()}
+            print("[OKX][DBG] ts", ts)
+            print("[OKX][DBG] pre-sign", {"method": method.upper(), "path": path, "body": body or ""})
+            print("[OKX][DBG] headers", safe)
+        return hdr
 
     def get_instrument(self, inst_id):
-        j = self.get("/api/v5/public/instruments", params={"instType":"SWAP"})
-        for it in j.get("data", []):
-            if it.get("instId") == inst_id: return it
+        r = self.rest.get("/api/v5/public/instruments?instType=SWAP")
+        r.raise_for_status()
+        for it in r.json()["data"]:
+            if it.get("instId") == inst_id:
+                return it
         raise RuntimeError(f"Instrument not found: {inst_id}")
 
+    def get_balance(self, ccy="USDT"):
+        path = "/api/v5/account/balance"
+        r = self.rest.get(path, headers=self._headers("GET", path, ""))
+        r.raise_for_status()
+        data = r.json()["data"][0]
+        for d in data.get("details", []):
+            if d.get("ccy") == ccy:
+                return float(d.get("eq", 0))
+        return float(data.get("totalEq", 0))
+
     def place_order(self, **kwargs):
-        j = self.post("/api/v5/trade/order", kwargs)
-        if j.get("code") != "0": raise RuntimeError(f"place_order error: {j}")
+        path = "/api/v5/trade/order"
+        payload = json.dumps(kwargs, separators=(",",":"))
+        r = self.rest.post(path, headers=self._headers("POST", path, payload), content=payload)
+        r.raise_for_status()
+        j = r.json()
+        if j.get("code") != "0":
+            raise RuntimeError(f"place_order error: {j}")
         return j["data"][0]
-
-    def cancel_order(self, **kwargs):
-        j = self.post("/api/v5/trade/cancel-order", kwargs)
-        if j.get("code") != "0": raise RuntimeError(f"cancel_order error: {j}")
-        return j["data"][0]
-
-    # low-level (used by attribution/replay)
-    def _get(self, path, params=None): return self.get(path, params)
-    def _post(self, path, payload): return self.post(path, payload)
